+++ xcode/Jasonette/Services/$agent/JasonAgentService.m
@@ -134,7 +134,11 @@
     NSString * summon = [raw stringByAppendingString:interface];
     webView.payload[@"state"] = @"rendered";
     [webView evaluateJavaScript:summon completionHandler:^(id _Nullable res, NSError * _Nullable error) {
-         NSLog(@"Injected");
+         DTLogInfo(@"Injected $agent into context");
+        if(error)
+        {
+            DTLogWarning(@"%@", error);
+        }
      }];
 
     // If there's a pending agent (because the method was called before the agent was initialized)
@@ -546,6 +550,7 @@
 - (void) inject:(NSString *)code into:(WKWebView *)agent {
     [agent evaluateJavaScript:code completionHandler:^(id _Nullable res, NSError * _Nullable error) {
          // Step 2. Execute the method with params
+         DTLogDebug(@"Injected code into agent");
          [[Jason client] success];
      }];
 }
@@ -585,7 +590,7 @@
         // Instead all communication back to Jasonette is taken care of by an explicit $agent.response() call
              if (error)
              {
-                 NSLog(@"%@", error);
+                 DTLogWarning(@"%@", error);
                  agent.payload[@"pending"] = options;
         // The agent might not be ready. Put it in a queue.
              }
@@ -604,7 +609,7 @@
     {
         [progressView setAlpha:1.0f];
         [progressView setProgress:((WKWebView *) object).estimatedProgress animated:YES];
-        NSLog(@"%f", progressView.progress);
+        DTLogDebug(@"%f", progressView.progress);
         if (((WKWebView *) object).estimatedProgress >= 1.0f)
         {
             [UIView animateWithDuration:0.3 delay:0.3 options:UIViewAnimationOptionCurveEaseOut animations:^{